+++ src/components/CodeBlock/LiveContainer.tsx
@@ -6,8 +6,8 @@ import { CssBaseLine } from 'smarthr-normalize-css';
 import * as ui from 'smarthr-ui';
 import styled, { StyleSheetManager, ThemeProvider, css } from 'styled-components';
 import { CSS_COLOR } from '@/constants/style';
-import styles from './LiveContainer.module.css';
-import sharedStyles from './shared.module.css';
+import styles from './LiveContainer.module.scss';
+import sharedStyles from './shared.module.scss';
 import ComponentPreview from '../ComponentPreview';
 import type { LiveContainerProps } from './';
 import CopyButton from './CopyButton';
