+++ src/components/Sidebar/RecursiveSidebarItem.astro
@@ -40,7 +40,7 @@ if (items.length === 0) {
   })
 }
 
-<style>
+<style lang="scss">
   .depth2Item {
     font-size: 18px;
     font-weight: bold;
