+++ src/components/article/ComponentCaptures.astro
@@ -31,7 +31,7 @@ const storyGroups = await fetchComponentCaptures();
   }
 </div>
 
-<style>
+<style lang="scss">
   .componentGroupSection {
     && {
       > h2 {
