+++ src/components/article/ComponentPropsTable/ComponentPropsTable.astro
@@ -57,7 +57,7 @@ const fragmentId = (propsName: string) => `props-${propsName.replace(' ', '-')}`
   )
 }
 
-<style>
+<style lang="scss">
   .wrapper {
     border: var(--border-shorthand);
     margin-top: 20px;
