+++ src/components/article/ComponentPropsTable/TypeTag.astro
@@ -34,7 +34,7 @@ const { content } = Astro.props;
   {content}
 </span>
 
-<style>
+<style lang="scss">
   span {
     font-size: var(--font-size-14);
     color: var(--color-white);
