+++ src/components/article/CustomLink.astro
@@ -17,7 +17,7 @@ const isExternal = href.match(/^https?:\/\/(?!smarthr\.design).*?$/) !== null;
   {isExternal && <FaExternalLinkAltIcon />}
 </a>
 
-<style>
+<style lang="scss">
   a {
     svg {
       margin-inline: 0.25em;
