+++ src/components/article/FragmentTitle.astro
@@ -20,7 +20,7 @@ const Wrapper = tag as keyof JSX.IntrinsicElements;
   </a>
 </Wrapper>
 
-<style>
+<style lang="scss">
   .wrapper {
     position: relative;
     scroll-margin-top: 80px;
