+++ src/components/article/ImgWithDesc.astro
@@ -9,7 +9,7 @@ type Props = {
   <figcaption>{Astro.props.description}</figcaption>
 </figure>
 
-<style>
+<style lang="scss">
   figure {
     display: block;
     margin: 0;
