+++ src/components/article/PageIndex/PageIndex.astro
@@ -37,7 +37,7 @@ subPages.sort((a, b) => (a.data.order && b.data.order ? a.data.order - b.data.or
   </div>
 </div>
 
-<style>
+<style lang="scss">
   .wrapper {
     margin-block: 4rem;
   }
