+++ src/components/article/VideoEmbed.astro
@@ -31,7 +31,7 @@ const { source, poster, title, youtubeUrl, tracks } = Astro.props;
   }
 </>
 
-<style>
+<style lang="scss">
   video {
     max-width: 100%;
     width: 100%;
